+++ src/redux/modules/asyncTodo/index.js
@@ -1,5 +1,5 @@
 import todo, { fetchTodos } from './asyncTodo'
-import { editTodo, addTodo, removeTodo } from './rest'
+import { editTodo, addTodo, removeTodo } from './others'
 
 export { fetchTodos, editTodo, addTodo, removeTodo }
 export default todo
